refactor(contracts): make receipts-audit remediation guide single-sourced
The Agent Receipts Audit remediation items (title/gap/fix, one per
question) were hand-duplicated between apps/marketing's content module and
apps/server's confirmation email, with a 'kept in sync by hand' comment.
That violated the fleet's single-source-of-truth and extend-before-create
rules, and the premise was wrong: apps/server already depends on
@revealui/contracts (it imports pricing, auth, and other contracts today).

Move the twelve items into @revealui/contracts/receipts-audit, following
the ARCHITECTURE_REVIEW_PRICE precedent in ./pricing.ts. Both apps import
the same array now:
  - apps/marketing/app/content/receipts-audit.ts keeps only the
    page-specific heading/intro wrap.
  - apps/server/src/routes/waitlist.ts deletes the local
    RECEIPTS_AUDIT_GUIDE copy and the sync-by-hand comment, importing
    RECEIPTS_AUDIT_REMEDIATION_ITEMS directly.

Added a drift-guard test asserting the email builder renders exactly the
twelve items the contracts module exports, and proved it (and the existing
pinned-title test) fail when the two diverge by truncating the email
builder's input to 11 items, then reverted.
